# ARROW-15136: [C++] Make S3FS tests faster